Prevent an unneeded warning message dialog that pops up for single file resources when the resource is configured to be read only and writeFile is called in aboutToQuit.
svn path=/trunk/KDE/kdepim/akonadi/; revision=990794

+++ resources/ical/icalresource.cpp
@@ -102,7 +102,9 @@ bool ICalResource::retrieveItem( const Akonadi::Item &item, const QSet<QByteArra
 
 void ICalResource::aboutToQuit()
 {
-  writeFile();
+  if ( !Settings::self()->readOnly() )
+    writeFile();
+
   Settings::self()->writeConfig();
 }
